Added classify_book.py
Added classify_book.py, which combines the two search methods into a
single program. It first prompts for the ISBN, and then if that fails,
prompts for the title and author. It will save to two separate CSV
files. In addition, did a little bit of cleanup of my other programs.